Provider rest tests
