+++ cfme/tests/infrastructure/test_rest_providers.py
@@ -1,0 +1,118 @@
+import fauxfactory
+import pytest
+
+from cfme.configure.configuration import server_roles_disabled
+from cfme.rest import a_provider as _a_provider
+from utils.virtual_machines import deploy_template
+from utils.version import current_version
+from utils.wait import wait_for
+
+
+@pytest.fixture(scope='module')
+def a_provider():
+    return _a_provider()
+
+
+def test_provider_refresh(request, a_provider, rest_api):
+    """Test checking that refresh invoked from the REST API works.
+    It provisions a VM when the Provider inventory functionality is disabled, then the functionality
+    is enabled and we wait for refresh to finish by checking the field in provider and then we check
+    whether the VM appeared in the provider.
+    Prerequisities:
+        * A provider that is set up, with templates suitable for provisioning.
+    Steps:
+        * Disable the ``ems_inventory`` and ``ems_operations`` roles
+        * Provision a VM
+        * Store old refresh date from the provider
+        * Initiate refresh
+        * Wait until the refresh date updates
+        * The VM should appear soon.
+    Metadata:
+        test_flag: rest
+    """
+    if "refresh" not in rest_api.collections.providers.action.all:
+        pytest.skip("Refresh action is not implemented in this version")
+    provider_rest = rest_api.collections.providers.get(name=a_provider.name)
+    with server_roles_disabled("ems_inventory", "ems_operations"):
+        vm_name = deploy_template(
+            a_provider.key,
+            "test_rest_prov_refresh_{}".format(fauxfactory.gen_alphanumeric(length=4)))
+        request.addfinalizer(lambda: a_provider.mgmt.delete_vm(vm_name))
+    provider_rest.reload()
+    old_refresh_dt = provider_rest.last_refresh_date
+    assert provider_rest.action.refresh()["success"], "Refresh was unsuccessful"
+    wait_for(
+        lambda: provider_rest.last_refresh_date,
+        fail_func=provider_rest.reload,
+        fail_condition=lambda refresh_date: refresh_date == old_refresh_dt,
+        num_sec=720,
+        delay=5,
+    )
+    # We suppose that thanks to the random string, there will be only one such VM
+    wait_for(
+        lambda: len(rest_api.collections.vms.find_by(name=vm_name)),
+        fail_condition=lambda l: l == 0,
+        num_sec=180,
+        delay=10,
+    )
+    vm = rest_api.collections.vms.get(name=vm_name)
+    if "delete" in vm.action.all:
+        vm.action.delete()
+
+
+def test_provider_edit(request, a_provider, rest_api):
+    """Test editing a provider using REST API.
+    Prerequisities:
+        * A provider that is set up. Can be a dummy one.
+    Steps:
+        * Retrieve list of providers using GET /api/providers , pick the first one
+        * POST /api/providers/<id> (method ``edit``) -> {"name": <random name>}
+        * Query the provider again. The name should be set.
+    Metadata:
+        test_flag: rest
+    """
+    if "edit" not in rest_api.collections.providers.action.all:
+        pytest.skip("Refresh action is not implemented in this version")
+    provider_rest = rest_api.collections.providers[0]
+    new_name = fauxfactory.gen_alphanumeric()
+    old_name = provider_rest.name
+    request.addfinalizer(lambda: provider_rest.action.edit(name=old_name))
+    provider_rest.action.edit(name=new_name)
+    provider_rest.reload()
+    assert provider_rest.name == new_name
+
+
+@pytest.mark.parametrize(
+    "from_detail", [True, False],
+    ids=["delete_from_detail", "delete_from_collection"])
+def test_provider_crud(request, rest_api, from_detail):
+    """Test the CRUD on provider using REST API.
+    Steps:
+        * POST /api/providers (method ``create``) <- {"hostname":..., "name":..., "type":
+            "EmsVmware"}
+        * Remember the provider ID.
+        * Delete it either way:
+            * DELETE /api/providers/<id>
+            * POST /api/providers (method ``delete``) <- list of dicts containing hrefs to the
+                providers, in this case just list with one dict.
+    Metadata:
+        test_flag: rest
+    """
+    if "create" not in rest_api.collections.providers.action.all:
+        pytest.skip("Create action is not implemented in this version")
+
+    if current_version() < "5.5":
+        provider_type = "EmsVmware"
+    else:
+        provider_type = "ManageIQ::Providers::Vmware::InfraManager"
+    provider = rest_api.collections.providers.action.create(
+        hostname=fauxfactory.gen_alphanumeric(),
+        name=fauxfactory.gen_alphanumeric(),
+        type=provider_type,
+    )[0]
+    if from_detail:
+        provider.action.delete()
+        provider.wait_not_exists(num_sec=30, delay=0.5)
+    else:
+        rest_api.collections.providers.action.delete(provider)
+        provider.wait_not_exists(num_sec=30, delay=0.5)
